Refactor: removed IsometricGameBoard, IsometricGrid, IsometricTile; integrated tile rendering logic into Tile component for improved maintainability and performance.
